pyproject.toml
[tool.poetry]
name = "qpuc-tieng-edition"
version = "0.1.0"
description = ""
authors = ["quenchhe <71766652+quenchhe@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
pygame = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
